Adding directory to IaC test config
Some customers want to point Snyk IaC to a directory within their codebase, and we support that use case with the Snyk CLI, but it does not work for this GitHub Action. Therefore, adding a directory field may work and provide us a way to add to the original use case.